🖥️ ui(web): add ShinyText component to StarView for indicating recently unlocked content
