+++ apps/web/src/ui/space/widgets/pages/Space/Planet/Star/AnimatedImage/index.tsx
@@ -1,35 +1,3 @@
-import { motion, type Variants } from 'motion/react'
+import { AnimatedImageView } from './AnimatedImageView'
 
-import type { ReactNode } from 'react'
-
-const variants: Variants = {
-  default: {
-    scale: 1,
-  },
-  pulse: {
-    scale: 1.15,
-    transition: {
-      repeat: Infinity,
-      repeatType: 'mirror',
-      duration: 0.4,
-    },
-  },
-}
-
-type AnimatedImageProps = {
-  children: ReactNode
-  shouldAnimate: boolean
-}
-
-export function AnimatedImage({ children, shouldAnimate }: AnimatedImageProps) {
-  return (
-    <motion.div
-      variants={variants}
-      initial='default'
-      animate={shouldAnimate ? 'pulse' : ''}
-      className='-ml-[36px]'
-    >
-      {children}
-    </motion.div>
-  )
-}
+export const AnimatedImage = AnimatedImageView
