+++ apps/web/src/ui/space/widgets/pages/Space/Planet/Star/AnimatedRocket/AnimatedRocketView.tsx
@@ -1,0 +1,44 @@
+import Image from 'next/image'
+import { motion, type Variants } from 'motion/react'
+
+import { useAuthContext } from '@/ui/auth/contexts/AuthContext'
+import { useImage } from '@/ui/global/hooks/useImage'
+
+const rocketVariants: Variants = {
+  hidden: {
+    opacity: 0,
+    y: '-100vh',
+  },
+  visible: {
+    opacity: 1,
+    y: 0,
+    transition: {
+      duration: 1.5,
+      type: 'spring',
+    },
+  },
+}
+
+type Props = {
+  shouldAnimate: boolean
+}
+
+export const AnimatedRocketView = ({ shouldAnimate }: Props) => {
+  const { user } = useAuthContext()
+  const rocketImage = user ? useImage('rockets', user.rocket.image.value) : ''
+
+  return (
+    <motion.div
+      variants={rocketVariants}
+      initial='hidden'
+      animate='visible'
+      className='h-20 w-20'
+    >
+      {shouldAnimate && user && (
+        <div className='relative h-20 w-20 rotate-180'>
+          <Image src={rocketImage} alt={`Foguete ${user.rocket.name.value}`} fill />
+        </div>
+      )}
+    </motion.div>
+  )
+}
