+++ apps/web/src/ui/space/widgets/pages/Space/Planet/Star/index.tsx
@@ -1,49 +1,47 @@
 'use client'
 
-import { Id, Slug } from '@stardust/core/global/structures'
+import { Id, Name, OrdinalNumber, Slug } from '@stardust/core/global/structures'
 
 import type { AnimationRef } from '@/ui/global/widgets/components/Animation/types'
 import { useRef } from 'react'
 import { useStar } from './useStar'
 import { useRest } from '@/ui/global/hooks/useRest'
 import { StarView } from './StarView'
+import { useAuthContext } from '@/ui/global/hooks/useAuthContext'
+import { useSpaceContext } from '@/ui/space/contexts/SpaceContext'
 
 type StarProps = {
-  id: string
-  name: string
-  slug: string
-  number: number
-  isUnlocked: boolean
-  isLastUnlockedStar: boolean
+  id: Id
+  name: Name
+  slug: Slug
+  number: OrdinalNumber
 }
 
-export const Star = ({
-  id,
-  name,
-  number,
-  isUnlocked,
-  slug,
-  isLastUnlockedStar,
-}: StarProps) => {
+export const Star = ({ id, name, number, slug }: StarProps) => {
   const { challengingService } = useRest()
+  const { user } = useAuthContext()
+  const { lastUnlockedStarId } = useSpaceContext()
+  const isLastUnlockedStar = lastUnlockedStarId === id.value
   const starAnimationRef = useRef<AnimationRef>(null)
   const { lastUnlockedStarRef, handleStarClick } = useStar({
-    starId: Id.create(id),
-    starSlug: Slug.create(slug),
+    starId: id,
+    starSlug: slug,
     isLastUnlockedStar,
     starAnimationRef,
     challengingService,
   })
 
-  return (
-    <StarView
-      name={name}
-      number={number}
-      isUnlocked={isUnlocked}
-      isLastUnlockedStar={isLastUnlockedStar}
-      lastUnlockedStarRef={lastUnlockedStarRef}
-      starAnimationRef={starAnimationRef}
-      onClick={handleStarClick}
-    />
-  )
+  if (user)
+    return (
+      <StarView
+        name={name.value}
+        number={number.value}
+        isUnlocked={user?.hasUnlockedStar(id).isTrue}
+        isRecentlyUnlocked={user?.hasRecentlyUnlockedStar(id).isTrue}
+        isLastUnlockedStar={isLastUnlockedStar}
+        lastUnlockedStarRef={lastUnlockedStarRef}
+        starAnimationRef={starAnimationRef}
+        onClick={handleStarClick}
+      />
+    )
 }
